git-svn-id: svn://123.57.173.36/meijia-parent@1217 52ac4245-7803-4e87-b018-04b7c5ce731e

+++ simi-dao/src/main/java/com/simi/po/dao/xcloud/XcompanyCheckinMapper.java
@@ -10,7 +10,7 @@ public interface XcompanyCheckinMapper {
 
     int insert(XcompanyCheckin record);
 
-    int insertSelective(XcompanyCheckin record);
+    Long insertSelective(XcompanyCheckin record);
 
     XcompanyCheckin selectByPrimaryKey(Long id);
